+++ artifacts/contracts/MintNFT.sol/MintNFT.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "../../build-info/54883b94ed30d40759379079cffe2a90.json"
+  "buildInfo": "../../build-info/771f88177e9bea12d495b15e71167c9c.json"
 }
