+++ utils/verify.js
@@ -1,22 +1,23 @@
-const { run, network } = require('hardhat')
-const { networkConfig } = require('../helper-hardhat-config')
+const { run, network } = require("hardhat")
+const { networkConfig } = require("../helper-hardhat-config")
 
 const verify = async (contractAddress, args) => {
-  console.log('Verifying contract...')
-  try {
-    await run('verify:verify', {
-      address: contractAddress,
-      constructorArguments: args,
-    })
-  } catch (e) {
-    if (e.message.toLowerCase().includes('already verified')) {
-      console.log('Already verified!')
-    } else {
-      console.log(e)
+    console.log("Verifying contract...")
+    try {
+        await run("verify:verify", {
+            address: contractAddress,
+            constructorArguments: args,
+            blockConfirmation: 6,
+        })
+    } catch (e) {
+        if (e.message.toLowerCase().includes("already verified")) {
+            console.log("Already verified!")
+        } else {
+            console.log(e)
+        }
     }
-  }
 }
 
 module.exports = {
-  verify,
+    verify,
 }
